requirements.txt
fastapi>=0.135,<1.0
uvicorn>=0.42,<1.0
httpx>=0.28,<1.0
psycopg2-binary>=2.9,<3.0
python-jose>=3.3,<4.0
pwdlib>=0.2,<1.0
argon2-cffi>=23.1,<24.0
python-multipart>=0.0.9,<1.0
email-validator>=2.0,<3.0
